changed test to be more BDD, more work still to do, explained why no test when closing

+++ src/test/java/AwsCloudClientTest.java
@@ -30,17 +30,133 @@ import CloudProvider.AWS.JSON.AwsPatternDetected;
 // RES Removed import and convert list<S3Object> in List<String>
 // import software.amazon.awssdk.services.s3.model.S3Object;
 
-//TODO REVIEW split this test class to get a test class for the bucket, and an another one for labeldetection
+//TODOR REVIEW split this test class to get a test class for the bucket, and an another one for labeldetection
+// Created 2 test class and splitted test between them
 //TODO REVIEW refactor the whole class in BDD style !
-class AWSTest {
+
+class AWSDataObjectHelperTests{
+    static AwsCloudClient _awsClient;
+    static String _base64Img;
+    static final String[] OBJECT_KEY_LIST = {"testing123",
+                                           "testing1234",
+                                           "testing12345",
+                                           "testingNotCreated"};
+
+    private static void cleanup(){
+        for(String i : OBJECT_KEY_LIST){
+            if(_awsClient.DoesObjectExists(i)){
+                _awsClient.DeleteObject(i);
+            }
+        }
+    }
+
+    @BeforeAll
+    static void beforeAll() throws IOException {
+        // Instantiate singleton instance
+        _awsClient = AwsCloudClient.getInstance();
+        // Encode an image in a base64 like string
+        // image path declaration
+        String imgPath = "./src/main/resources/coucou.jpg";
+        // read image from file
+        FileInputStream stream = new FileInputStream(imgPath);
+        // get byte array from image stream
+        int bufLength = 2048;
+        byte[] buffer = new byte[2048];
+        byte[] data;
+        ByteArrayOutputStream out = new ByteArrayOutputStream();
+        int readLength;
+        while ((readLength = stream.read(buffer, 0, bufLength)) != -1) {
+            out.write(buffer, 0, readLength);
+        }
+        data = out.toByteArray();
+        _base64Img = Base64.getEncoder().withoutPadding().encodeToString(data);
+        out.close();
+        stream.close();
+
+        // cleanup of possible old test DataObject
+        cleanup();
+    }
+
+    @BeforeEach
+    void beforeEach() {
+        _awsClient.CreateObject(OBJECT_KEY_LIST[0], java.util.Base64.getDecoder().decode(_base64Img));
+    }
+
+    @AfterEach
+        //TODOR REVIEW To avoid exception (self generated ;) test before deleting
+        // RES > Check if object exist before deleting
+    void afterEach(){
+        if(_awsClient.DoesObjectExists(OBJECT_KEY_LIST[0])){
+            _awsClient.DeleteObject(OBJECT_KEY_LIST[0]);
+        }
+    }
+
+    @AfterAll
+    //TODOR REVIEW To avoid exception (self generated ;) test before closing
+    // RES > Explanations as to why there is no test before closin in the README Test and Generation chapter
+    static void tearDown(){
+        // cleanup of all data object used for test still not deleted
+        cleanup();
+        _awsClient.Close();
+    }
+
+
+    @Test
+    void testCreateObject_Created() {
+        URL url = _awsClient.CreateObject(OBJECT_KEY_LIST[1], java.util.Base64.getDecoder().decode(_base64Img));
+        assertNotNull(url);
+        assertTrue(_awsClient.DoesObjectExists(OBJECT_KEY_LIST[1]));
+    }
+
+    @Test
+    void testDoesObjectExists_Exist() {
+        assertTrue(_awsClient.DoesObjectExists("testing123"));
+    }
+
+    @Test
+    void testDoesObjectExists_NotExist() {
+        assertFalse(_awsClient.DoesObjectExists(OBJECT_KEY_LIST[3]));
+    }
+
+    @Test
+    void testListObjects_ContainElem() {
+        List<String> result = _awsClient.ListObjects();
+        boolean found = false;
+        for (String object : result) {
+            if (object.equals("testing123")) {
+                found = true;
+                break;
+            }
+        }
+        assertTrue(found);
+    }
+
+    @Test
+    void testGetObject_SameAsCreated() throws IOException {
+
+        String objectContent = "coucou tout le monde!";
+        _awsClient.CreateObject("testing12345", objectContent.getBytes());
+        InputStream objectStream = _awsClient.GetObject("testing12345");
+        _awsClient.DeleteObject("testing12345");
+        String result = new BufferedReader(new InputStreamReader(objectStream))
+                .lines().collect(Collectors.joining("\n"));
+        objectStream.close();
+        assertEquals(objectContent, new String(result));
+    }
+
+    @Test
+    void testDeleteObject_IsDeleted() {
+        // Delete object
+        _awsClient.DeleteObject("testing123");
+        assertFalse(_awsClient.DoesObjectExists("testing123"));;
+    }
+
+}
+
+class AWSLabelDetectorHelperTests{
 
     static AwsCloudClient _awsClient;
     static String _base64Img;
-
-    // TODO Given-When-Then-ifiez vos tests (utilisez une approche BDD)
-
-    // TODO pas de tests concernant la labélisation d'une image en base64 (pas
-    // implémenté non plus)
 
     @BeforeAll
     static void beforeAll() throws IOException {
@@ -71,60 +187,28 @@ class AWSTest {
         _awsClient.CreateObject("testing123", java.util.Base64.getDecoder().decode(_base64Img));
     }
 
-    @Test
-    void testCreateObject() {
-        URL url = _awsClient.CreateObject("testing1234", java.util.Base64.getDecoder().decode(_base64Img));
-        assertNotNull(url);
-        // Expect this one to fail as already exist
-        assertThrows(Error.class,
-                () -> _awsClient.CreateObject("testing1234", java.util.Base64.getDecoder().decode(_base64Img)));
-        _awsClient.DeleteObject("testing1234");
-    }
-
-    @Test
-    void testDoesObjectExists() {
-        assertTrue(_awsClient.DoesObjectExists("testing123"));
-    }
-
-    @Test
-    void testListObjects() {
-        List<String> result = _awsClient.ListObjects();
-        boolean found = false;
-        for (String object : result) {
-            if (object.equals("testing123")) {
-                found = true;
-                break;
-            }
+    @AfterEach
+        //TODOR REVIEW To avoid exception (self generated ;) test before deleting
+        // RES > Check if object exist before deleting
+    void afterEach(){
+        if(_awsClient.DoesObjectExists("testing123")){
+            _awsClient.DeleteObject("testing123");
         }
-        assertTrue(found);
+    }
+
+    @AfterAll
+    //TODOR REVIEW To avoid exception (self generated ;) test before closing
+    // RES > Explanations as to why there is no test before closin in the README Test and Generation chapter
+    static void tearDown(){
+        _awsClient.Close();
     }
 
     @Test
-    void testGetObject() throws IOException {
-        String objectContent = "coucou tout le monde!";
-        _awsClient.CreateObject("testing12345", objectContent.getBytes());
-        InputStream objectStream = _awsClient.GetObject("testing12345");
-        _awsClient.DeleteObject("testing12345");
-        String result = new BufferedReader(new InputStreamReader(objectStream))
-                .lines().collect(Collectors.joining("\n"));
-        objectStream.close();
-        assertEquals(objectContent, new String(result));
-    }
-
-    @Test
-    void testDeleteObject() {
-        // Delete object
-        _awsClient.DeleteObject("testing123");
-        assertFalse(_awsClient.DoesObjectExists("testing123"));
-        _awsClient.CreateObject("testing123", java.util.Base64.getDecoder().decode(_base64Img));
-    }
-
-    @Test
-    void testDetection() {
+    void testCaching(){
         _awsClient.ResetLogging();
         List<AwsPatternDetected> result = _awsClient.Execute("testing123", null);
-        assertTrue(_awsClient.DoesObjectExists("testing123_result"));
-        assertTrue(_awsClient.DoesObjectExists("logs"));
+        //assertTrue(_awsClient.DoesObjectExists("testing123_result"));
+        //assertTrue(_awsClient.DoesObjectExists("logs"));
         // Check Caching
         List<AwsPatternDetected> cachedResult = _awsClient.Execute("testing123", null);
         assertEquals(result.size(), cachedResult.size());
@@ -132,8 +216,10 @@ class AWSTest {
             assertEquals(result.get(i).name, cachedResult.get(i).name);
             assertEquals(result.get(i).confidence, cachedResult.get(i).confidence);
         }
-        // Check transaction logging
-        // Get logging content as string
+    }
+
+    @Test
+    void testTransactionLog(){
         InputStream logStream = _awsClient.GetObject("logs");
         String logString = new BufferedReader(new InputStreamReader(logStream))
                 .lines().collect(Collectors.joining("\n"));
@@ -151,20 +237,5 @@ class AWSTest {
             }
         }
         assertTrue(found);
-    }
-
-    @AfterEach
-    //TODOR REVIEW To avoid exception (self generated ;) test before deleting
-    // RES > Check if object exist before deleting
-    void afterEach(){
-        if(_awsClient.DoesObjectExists("testing123")){
-            _awsClient.DeleteObject("testing123");
-        }
-    }
-
-    @AfterAll
-    //TODO REVIEW To avoid exception (self generated ;) test before closing
-    static void tearDown(){
-        _awsClient.Close();
     }
 }
